requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
dj-database-url==2.0.0
Django==4.2.3
django-appconf==1.0.5
django-extensions==3.2.3
django-grappelli==3.0.6
django-oauth-toolkit==2.3.0
django-user-accounts==3.2.0
idna==3.4
jwcrypto==1.5.0
oauth2-provider==0.0
oauthlib==3.2.2
psycopg2==2.9.6
pycparser==2.21
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.3
wrapt==1.15.0
Pillow==10.0.0
